+++ src/cmake/modules/FindPugiXML.cmake
@@ -21,6 +21,14 @@ FIND_PACKAGE_HANDLE_STANDARD_ARGS (PugiXML DEFAULT_MSG PUGIXML_LIBRARY
 if (PUGIXML_FOUND)
     set (PUGIXML_INCLUDES ${PUGIXML_INCLUDE_DIR})
     set (PUGIXML_LIBRARIES ${PUGIXML_LIBRARY})
+
+    if (NOT TARGET pugixml)
+        add_library(pugixml UNKNOWN IMPORTED)
+        set_target_properties(pugixml PROPERTIES
+            INTERFACE_INCLUDE_DIRECTORIES "${PUGIXML_INCLUDES}")
+        set_property(TARGET pugixml APPEND PROPERTY
+            IMPORTED_LOCATION "${PUGIXML_LIBRARIES}")
+    endif ()
 endif()
 
 mark_as_advanced (PUGIXML_LIBRARY PUGIXML_INCLUDE_DIR)
