+++ src/include/OpenImageIO/detail/pugixml/pugixml.cpp
