+++ src/libOpenImageIO/xmp.cpp
@@ -19,7 +19,7 @@ extern "C" {
 #if USE_EXTERNAL_PUGIXML
 #    include <pugixml.hpp>
 #else
-#    include <OpenImageIO/pugixml.hpp>
+#    include <OpenImageIO/detail/pugixml/pugixml.hpp>
 #endif
 
 #define DEBUG_XMP_READ 0
